bizland/requirement.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-js-asset==2.2.0
django-modeltranslation==0.19.11
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.1
tzdata==2024.2
